added login function and prepares the frontend
