+++ app/Http/routes.php
@@ -1,6 +1,11 @@
 <?php
 
-Route::resource('section', 'SectionController');
-Route::resource('document', 'DocumentController');
-Route::resource('user', 'UserController');
-Route::post('/login', 'Auth\AuthController@login');
+Route::group(['namespace' => 'Backend', 'prefix' => 'api', 'middleware' => 'auth'], function() {
+	Route::resource('section', 'SectionController');
+	Route::resource('document', 'DocumentController');
+	Route::resource('user', 'UserController');
+	Route::resource('project', 'ProjectController');
+});
+
+Route::post('login', 'AuthController@login');
+Route::get('logout', 'AuthController@logout');
