+++ composer.json
@@ -13,8 +13,7 @@
         "mockery/mockery": "0.9.*",
         "phpunit/phpunit": "~4.0",
         "phpspec/phpspec": "~2.1",
-        "doctrine/dbal": "2.4",
-        "catalyst/sentry": "dev-feature/laravel-5"
+        "doctrine/dbal": "2.4"
     },
     "autoload": {
         "classmap": [
